First working version of PLY IO with hlists
A lot happened in this PR, including a lot of experimenting. The
important part is that this is a proof of concept for PlyWriter
working with hlists to store the properties. This means there
is no dynamic dispatch and the compiler has perfect knowledge
of all types involved. So it should be able to optimize it quite
well.

The code will be vastly improved in future commits.