Change `io::Error` layout to a simple `Box<ErrorKind>`
The previous `Error` (now `ErrorKind`) was 56 bytes large. Since we
returned it from many functions, including very small functions,
reading and writing this return type took a lot of time. Since the
error should rarely actually exist, it's fine to box it. That way we
reduce the size to 8 bytes. That means we don't destroy all of our
stack frames.

In the PLY read benchmarks, the performance improved from 8% to 30%
(most around 15%).