Remove "type wishes" from `MemSink`
The whole concept was a nice idea, but it's just not worth it in
practice. For one, the Rust feature doesn't even work properly right
now. But more than that, even if it would work, type wishes make the
`StreamSource` code so much more complicated and verbose.

Let's take a look at all sources that could benefit from it:
- ASCII files where a value can be parsed as `f32` or `f64`
- Algorithms generating mesh data (like shapes or marching cubes)

For one, it's super unlikely that the extra precision of `f64` is
needed in the mesh processing domain. But assuming it is, it's more
viable to add a type parameter to the source and configure the source
via that parameter.

So in summary, type wishes: make the lox API more verbose, make source
implementation more complex and verbose, are not used in almost all
cases, and are not strictly necessary because there are other ways to
achieve the same effect. So for now, this library will go with the
more non-overengineered way and remove this feature.