Change default cast mode for `derive(Mem*)` to "lossy"
I think this is what people want more than 99% of the time. And is
there really a risk to having the cast mode "lossy" by default? Yes,
there might be some unexpected bad things, but the same is true for
the other way around: having "lossless" or "none" as default cast mode
often leads to meshes not being able to deal with input data. It
happened to me already a couple of times and each time I just thought
"oh, I forgot to relax the casting mode".