Redesign adjacency traits (merge existing ones together)
Having one trait per one of the 8 useful adjacency information turned
out to have several disadvantages. For one, it was very verbose and
the names weren't pretty.

More importantly, certain features didn't really fit into a specific
trait: for example, asking whether or not a vertex/face is on the
boundary. How do you do that? If you have F2F adjacency, you can use
that to answer the question for the face. In any case, to answer that
question, you need some kind of adjacency information. But since not
a single adjacency information would be responsible for this query,
we would build two new traits: `IsVertexBoundary` and `IsFaceBoundary`.
This would only make the situation worse.

So to reduce the number of traits and to make things a bit easier
towards the end of my master thesis, the traits were simplified into
three traits:
- BasicAdj
- FullAdj
- EdgeAdj

This trait system actually abstracts very well over the data structures
that will be offered in this library. However, in the future, the
trait system might change again to allow more fine grained control.