Remove `ElementRefMut`
The type didn't really serve a purpose until now. It will be
reintroduced later, but in a different form. To be really useful, we
want an iterator that can yields this RefMuts. But this is currently
not possible and would even be unsafe since `RefMut` could give access
to its underlying mesh reference. So since implementing them well is
not trivial and not very important right now and since the current
version doesn't do anything useful, this commit removes it.