Replace boxed iterators in adj traits with ugly GAT workaround
This improves performance notably. In the sqrt3 benchmark, it was
improved from 80ms to 65ms. Sadly, this introduces a lot of noise. But
at least the code using this doesn't need to be changed. And we will
change this to a nice form again, once GATs or `impl Trait` in traits
lands.