Remove traits that are combinations of different capabilities (a test!)
This is an experimental commit that might get reverted later. There are
optional capabilities offered by different mesh data structures. Some
features (methods) require multiple of these capabilities (for example,
`flip_edge` requires mutability and exposed edges). Before this commit,
there was a trait for each capability combination. But this meant that
there are many traits which is a bit chaotic. With this commit, all
methods live in `Mesh` or `MeshMut`. All other traits are just marker
traits that enable methods in the main traits.

I'm not sure this is necessarily a better idea than the design before.
But it is worth a try.