Comment out all IO related code
I am sitting on this unfinished crate for many years now. I finally had
some motivation to look at this again and noticed that it's just a very
large library. In order to get at least something released, I decided
to first ignore all IO stuff as that brings a number of additional
challenges. It makes the library significantly less usable, but not
unusable outright. Some users only need a data structure without ever
doing IO.

I was unsure whether to delete the code (it's still in git of course)
or to comment out. For now I decided for commenting out.