Rewrite `cast` module to be smaller and work without specialization
I think this implementation is nicer, hopefully also reducing compile
times as there are fewer traits and significantly fewer impls. But it
does not quite have the power of the previous implementation which is
just due to the lack of specialization. The previous `try_cast` could
be called with any type as there was a blanket impl for `TryCast`. This
cannot be achieved without specialization, so `try_cast` requires
`CastFrom`.