Remove `#![feature(associated_type_bounds)]`
With this, `lox` itself compiles on stable. However `lina` still 
requires nightly.

Note that associated type bounds are not just a syntax sugar thing but
enable truly new things, especially related to super trait bounds. For
example:

    trait Foo: Bar<A: Baz>
    trait Foo: Bar where Bar::A: Baz

Those two are not equivalent! Take a function saying `<T: Foo>`. In the
first definition the function is now guaranteed that `T` implements
`Bar` and that `<T as Bar>::A` implements `Baz`. In the second trait
definition the function itself has to prove (by adding another bound)
that `<T as Bar>::A` implements `Baz`!

This is cool but is one reason why there has hardly been any progress
on stabilizing this. It's non-trivial type system stuff. I unfortunately
don't expect it to be stabilized anytime soon. So out it goes.

This changes the semantics of the `Cast` trait in `util` due to the
super trait stuff explained above. This makes the trait less useful,
unfortunately.