Remove `TinyMap` and `TinySet` (for now)
The implementation is not optimal. Using stack space to store items is
the most important aspect of tiny map, but it should switch to a proper
map if the stack space is not sufficient instead of going with a O(n)
Vec-based solution.

So this has to be rewritten. Until then, it is removed.